Added ability to exclude characters (#21)
* * Added ability to exclude characters
* Added ability to exclude categories
* Refactored the `RangedSubsetBuilderImpl` to achieve the above.
* Created new concrete `RangedSubsetWithCategoriesImpl` and `RangedSubsetWithCategoriesAndExcludesImpl` classes to cater to categories and excludes.
* Removed the categories code from the `TypeParserImpl`
* Started to add more unit tests

* * Added more unit tests

* * Added more unit tests

* * Added more unit tests

* * Added more unit tests
* Improved test coverage
* Fixed code-smell as reported by Sonar

+++ easy-type-language-generator/src/main/java/land/artli/easytype/generator/language/CodePointRange.java
@@ -11,7 +11,7 @@ class CodePointRange {
   }
 
   static CodePointRange range(char fromChar, char toChar) {
-    return new CodePointRange((int) fromChar, (int) toChar);
+    return new CodePointRange(fromChar, toChar);
   }
 
   static CodePointRange range(int fromCodePoint, int toCodePoint) {
